requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
mysql-connector-python==8.2.0
Pillow==10.2.0
twilio==9.4.6
django-anymail==10.2
# Remove old email services
# sendgrid==6.10.0
# sib-api-v3-sdk==7.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-allauth==0.57.0
social-auth-app-django==5.2.0
django-axes==6.1.1
django-otp==1.2.2
argon2-cffi==23.1.0
whitenoise==6.5.0
dj-database-url==2.1.0
certifi==2023.7.22
django-storages==1.14.2
django-redis==5.4.0
boto3==1.28.38
sentry-sdk==1.30.0
